ai_roofing_archive/20250905T192517Z/calculate_materials_wrapper/requirements.txt
fastapi
uvicorn
httpx
google-auth
google-cloud-storage

requests
